feat(db): migrate from Neon Postgres to Cloudflare D1 + R2
Neon's free tier kept hitting its 5 GB/month network-transfer cap (402),
which made every page and the admin panel unreadable. Moving to Cloudflare's
own services removes that ceiling entirely: D1 is a Worker binding, so
queries never leave Cloudflare's network and there is no egress to meter.

Database (Neon Postgres -> D1/SQLite)
- src/db/schema.ts rewritten for drizzle sqlite-core; table and column names
  unchanged so every query in src/lib and src/app/api keeps working.
  serial->autoincrement, timestamptz->integer epoch ms, boolean->integer 0/1.
- src/db/index.ts uses the `DB` binding via getCloudflareContext(); the
  connection-pool helpers are gone (D1 has no pool).
- drizzle/0000_init_d1.sql is the canonical schema; src/db/d1-ddl.ts keeps an
  inline copy so a fresh database can still bootstrap itself.

PostgreSQL syntax removed from application code
- count(*)::int -> count(*) (10 sites)
- ilike -> like, and the local `like` bindings renamed so they stop shadowing
  the imported helper
- now() -> epoch ms passed in as a parameter
- is distinct from -> ifnull(...) <> ...
- db.execute() -> db.run()/db.all(); D1 has no execute() and returns rows
  under `results`, so SELECTs now use all() which yields a plain array

Media (R2)
- New src/lib/media-storage.ts writes uploads to the `UPLOADS` R2 bucket,
  falling back to public/ only when the binding is absent (plain next dev).
- /api/admin/upload stores to R2; /api/media/[...key] streams objects back.

Data
- Neon is still returning 402, so the local seed files were the only readable
  source. 7 categories / 92 posts / 159 products / 33 contents / 4 roles /
  15 settings / 3 messages were loaded into D1 ahead of time - the runtime
  seed would blow the Worker CPU budget on a 14 MB import.

+++ src/app/api/admin/sections/route.ts
@@ -57,7 +57,7 @@ export async function GET() {
           // therefore always returned 0 — resolve through `news_categories`
           // first, exactly like `/api/admin/posts` does.
           const result = await db
-            .select({ count: sql<number>`count(*)::int` })
+            .select({ count: sql<number>`count(*)` })
             .from(newsPosts)
             .innerJoin(newsCategories, eq(newsPosts.categoryId, newsCategories.id))
             .where(eq(newsCategories.sourceId, col.sourceId));
@@ -65,7 +65,7 @@ export async function GET() {
           continue;
         }
         if (col.dataSource === "inquiries") {
-          const result = await db.select({ count: sql<number>`count(*)::int` }).from(inquiries);
+          const result = await db.select({ count: sql<number>`count(*)` }).from(inquiries);
           col.itemCount = result[0]?.count ?? 0;
           continue;
         }
